feat: release v0.7.0 with corrected workloads and direct SQL transport
Use direct database connections with SSH reserved for host operations and
metrics. Correct Pagila and IMDb data distributions, temporal semantics,
query aggregates and selective indexes, and document profile size estimates.

Return newly created Pagila rentals through the open-inventory index and
serialize inventory moves with rental creation using a fresh snapshot after
the row lock. Protect JSON strings during legacy lshw structural recovery.
Add regression coverage for rental reuse, concurrent moves and JSON values.

Validation: 363 unit tests and 10 PostgreSQL 18.6 Pagila integration tests
passed; Ruff, wheel/sdist metadata and documentation checks, and installed
wheel validation passed. Chart tests require Node.js and were skipped locally.


skip-checks: true

+++ src/pg_perf_bench/workload_profiles/imdb/sql/select_13.sql
@@ -2,7 +2,7 @@ set search_path = 'imdb';
 
 BEGIN;
 SELECT MIN(mi.info) AS release_date,
-       MIN(miidx.info) AS rating,
+       MIN(miidx.info::numeric) AS rating,
        MIN(t.title) AS german_movie
 FROM company_name AS cn,
      company_type AS ct,
@@ -34,7 +34,7 @@ COMMIT;
 
 BEGIN;
 SELECT MIN(cn.name) AS producing_company,
-       MIN(miidx.info) AS rating,
+       MIN(miidx.info::numeric) AS rating,
        MIN(t.title) AS movie_about_winning
 FROM company_name AS cn,
      company_type AS ct,
@@ -69,7 +69,7 @@ COMMIT;
 
 BEGIN;
 SELECT MIN(cn.name) AS producing_company,
-       MIN(miidx.info) AS rating,
+       MIN(miidx.info::numeric) AS rating,
        MIN(t.title) AS movie_about_winning
 FROM company_name AS cn,
      company_type AS ct,
@@ -104,7 +104,7 @@ COMMIT;
 
 BEGIN;
 SELECT MIN(cn.name) AS producing_company,
-       MIN(miidx.info) AS rating,
+       MIN(miidx.info::numeric) AS rating,
        MIN(t.title) AS movie
 FROM company_name AS cn,
      company_type AS ct,
